Add availability notations and preproc macros to maintain swift 5.3, 5.4 support. Begin process of making Combine framework optional for non Apple platforms
feature/structured-concurrency
