+++ .github/workflows/ci.yml
@@ -14,10 +14,26 @@ jobs:
     environment: default
     strategy:
       matrix:
+        os:
+          - macos-11
+          - macos-12
         xcode:
+          - '12.4'
+          - '12.5.1'
           - '13.2'
           - '13.4'
           - '14.0'
+        exclude:
+          - os: macos-11
+            xcode: '13.2'
+          - os: macos-11
+            xcode: '13.4'
+          - os: macos-11
+            xcode: '14.0'
+          - os: macos-12
+            xcode: '12.4'
+          - os: macos-12
+            xcode: '12.5.1'
     steps:
       - uses: actions/checkout@v2
       - name: Select Xcode ${{ matrix.xcode }}
