+++ Package.swift
@@ -1,12 +1,25 @@
 // swift-tools-version:5.3
-// The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "NetworkService",
     platforms: [.iOS(.v13), .macOS(.v10_15), .tvOS(.v13), .watchOS(.v6)],
-    products: [
+    products: Product.products,
+    dependencies: [
+        .package(url: "https://github.com/AliSoftware/OHHTTPStubs.git", from: "9.1.0"),
+        .package(
+            name: "combine-schedulers",
+            url: "https://github.com/pointfreeco/combine-schedulers.git",
+            .upToNextMajor(from: "0.5.0")
+        ),
+    ],
+    targets: Target.targets
+)
+
+#if swift(>=5.5)
+extension Product {
+    static let products: [Product] = [
         .library(
             name: "NetworkService",
             targets: ["NetworkService"]
@@ -23,16 +36,11 @@ let package = Package(
             name: "NetworkServiceTestHelperAsyncBeta",
             targets: ["NetworkServiceTestHelperAsyncBeta"]
         ),
-    ],
-    dependencies: [
-        .package(url: "https://github.com/AliSoftware/OHHTTPStubs.git", from: "9.1.0"),
-        .package(
-            name: "combine-schedulers",
-            url: "https://github.com/pointfreeco/combine-schedulers.git",
-            .upToNextMajor(from: "0.5.0")
-        ),
-    ],
-    targets: [
+    ]
+}
+
+extension Target {
+    static let targets: [Target] = [
         .target(
             name: "NetworkService",
             dependencies: []
@@ -91,4 +99,54 @@ let package = Package(
             ]
         ),
     ]
-)
+}
+
+
+#else
+extension Product {
+    static let products: [Product] = [
+        .library(
+            name: "NetworkService",
+            targets: ["NetworkService"]
+        ),
+        .library(
+            name: "NetworkServiceTestHelper",
+            targets: ["NetworkServiceTestHelper"]
+        ),
+    ]
+}
+
+extension Target {
+    static let targets: [Target] = [
+        .target(
+            name: "NetworkService",
+            dependencies: []
+        ),
+        .testTarget(
+            name: "NetworkServiceTests",
+            dependencies: [
+                "NetworkService",
+                .product(name: "OHHTTPStubs", package: "OHHTTPStubs"),
+                .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
+            ]
+        ),
+        .target(
+            name: "NetworkServiceTestHelper",
+            dependencies: [
+                "NetworkService",
+                .product(name: "CombineSchedulers", package: "combine-schedulers"),
+            ],
+            swiftSettings: [
+                .unsafeFlags(["-Xfrontend", "-warn-concurrency"]),
+            ]
+        ),
+        .testTarget(
+            name: "NetworkServiceTestHelperTests",
+            dependencies: [
+                "NetworkServiceTestHelper",
+                .product(name: "CombineSchedulers", package: "combine-schedulers"),
+            ]
+        ),
+    ]
+}
+#endif
