+++ Sources/NetworkServiceAsyncBeta/NetworkService.swift
@@ -23,5 +23,5 @@ public final class NetworkService {
 }
 
 // MARK: NetworkService+NetworkServiceClient
-
+@available(swift 5.5)
 extension NetworkService: NetworkServiceClient {}
