+++ Sources/NetworkServiceAsyncBeta/NetworkServiceClient+Delete.swift
@@ -6,7 +6,6 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-import Combine
 import Foundation
 
 extension NetworkServiceClient {
@@ -20,12 +19,15 @@ extension NetworkServiceClient {
         _ url: URL,
         headers: [HTTPHeader] = []
     ) async -> Result<Data, Failure> {
-        var request = URLRequest(url: url)
-        request.method = .DELETE
-        headers.forEach { request.addValue($0) }
+        let request = URLRequest.service(url: url, headers: headers, method: .DELETE)
         return await start(request)
     }
+}
 
+#if canImport(Combine)
+import Combine
+
+extension NetworkServiceClient {
     /// Send a delete request to a `URL`
     /// - Parameters:
     ///   - url: The destination for the request
@@ -37,21 +39,22 @@ extension NetworkServiceClient {
         headers: [HTTPHeader] = [],
         decoder: Decoder
     ) async -> Result<ResponseBody, Failure>
-        where ResponseBody: Decodable, Decoder: TopLevelDecoder, Decoder.Input == Data
+    where ResponseBody: Decodable, Decoder: TopLevelDecoder, Decoder.Input == Data
     {
         await delete(url, headers: headers)
             .decode(with: decoder)
             .mapToNetworkError()
     }
-
+    
     /// Send a delete request to a `URL`
     /// - Parameters:
     ///     - url: The destination for the request
     ///     - headers: HTTP headers for the request
     /// - Returns: Type erased publisher with `TopLevelDecodable` output and `NetworkService`'s error domain for failure
     public func delete<ResponseBody>(_ url: URL, headers: [HTTPHeader] = []) async -> Result<ResponseBody, Failure>
-        where ResponseBody: TopLevelDecodable
+    where ResponseBody: TopLevelDecodable
     {
         await delete(url, headers: headers, decoder: ResponseBody.decoder)
     }
 }
+#endif
