+++ Sources/NetworkServiceAsyncBeta/NetworkServiceClient+Start.swift
@@ -10,30 +10,6 @@ import Combine
 import Foundation
 
 extension NetworkServiceClient {
-    // MARK: URLRequest
-
-    /// Start a `URLRequest`
-    /// - Parameter request: The request as a `URLRequest`
-    /// - Returns: Type erased publisher with decoded output and `NetworkService`'s error domain for failure
-    public func start<ResponseBody, Decoder>(
-        _ request: URLRequest,
-        with decoder: Decoder
-    ) async -> Result<ResponseBody, Failure>
-        where ResponseBody: Decodable, Decoder: TopLevelDecoder, Decoder.Input == Data
-    {
-        await start(request)
-            .decode(with: decoder)
-            .mapToNetworkError()
-    }
-
-    /// Start a `URLRequest`
-    /// - Parameter request: The request as a `URLRequest`
-    /// - Returns: Type erased publisher with decoded output and `NetworkService`'s error domain for failure
-    public func start<ResponseBody>(_ request: URLRequest) async -> Result<ResponseBody, Failure>
-        where ResponseBody: TopLevelDecodable
-    {
-        await start(request, with: ResponseBody.decoder)
-    }
 
     /// Start a `URLRequest`
     /// - Parameter request: The request as a `URLRequest`
@@ -87,3 +63,30 @@ extension NetworkServiceClient {
         }
     }
 }
+
+#if canImport(Combine)
+extension NetworkServiceClient {
+    /// Start a `URLRequest`
+    /// - Parameter request: The request as a `URLRequest`
+    /// - Returns: Type erased publisher with decoded output and `NetworkService`'s error domain for failure
+    public func start<ResponseBody, Decoder>(
+        _ request: URLRequest,
+        with decoder: Decoder
+    ) async -> Result<ResponseBody, Failure>
+        where ResponseBody: Decodable, Decoder: TopLevelDecoder, Decoder.Input == Data
+    {
+        await start(request)
+            .decode(with: decoder)
+            .mapToNetworkError()
+    }
+
+    /// Start a `URLRequest`
+    /// - Parameter request: The request as a `URLRequest`
+    /// - Returns: Type erased publisher with decoded output and `NetworkService`'s error domain for failure
+    public func start<ResponseBody>(_ request: URLRequest) async -> Result<ResponseBody, Failure>
+        where ResponseBody: TopLevelDecodable
+    {
+        await start(request, with: ResponseBody.decoder)
+    }
+}
+#endif
